feat: add support to create forked blockchains
This commits adds the foundation to generate forked blockchains.
The generate_fork.py script first generates a "shared" chain and uses this as a baseline to generate a "main" and a "forked" chain.
Currently both generated chains (main and forked) are identical. The runner pipeline needs to be modified for one/both of the chains to create a realistic fork.
Main changes: add node directory setting, launch bitcoind with -reindex option and set the correct Runner.current_time for pipelines that build upon another chain.